fix double counting bug in statistics overview
This fixes a bug in the statistic calculation where filament length,
weight and some other entries were doubled.
As already assumed by donlook in #161 this is due to the existence of
print job entries for every tool and a total entry. This simple fix
simply skips the total entries, only including the ones per tool.

Fixes #161